requirements.txt
alembic==1.0.9
asn1crypto==0.24.0
async-generator==1.10
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
decorator==4.4.0
Django==1.11.20
django-contrib-comments==1.9.1
filebrowser-safe==0.5.0
future==0.17.1
grappelli-safe==0.5.2
idna==2.8
ipython-genutils==0.2.0
Jinja2==2.10.1
jupyterhub==0.9.6
Mako==1.0.9
MarkupSafe==1.1.1
Mezzanine==4.3.1
oauthlib==3.0.1
pamela==1.0.0
Pillow==6.0.0
prometheus-client==0.6.0
pycparser==2.19
pycurl==7.43.0.2
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
python-editor==1.0.4
python-oauth2==1.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
sqlparse==0.3.0
tornado==6.0.2
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.24.2
webencodings==0.5.1
